feature(onboarding): refactoring
